consolidate /i/v0/e/ and /batch/ suites again on new handling path
